Split dom events registration
